kantan/requirements.txt
Django==3.2.14
django-cors-headers==3.13.0
django-extensions==3.2.0
django-translatable-fields==0.0.1
django-translations==1.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.3
drf-yasg==1.21.3
django-jazzmin==2.5.0
